+++ LecturaYCalculos/src/modelo/Palabra.java
@@ -3,14 +3,14 @@ package modelo;
 public class Palabra {
 public String palabra;
 public int repeticiones;
-public float cantInfo,entropia;
+public Double cantInfo,entropia;
 
 public Palabra() {
 	super();
 	this.palabra = null;
 	this.repeticiones = 0;
-	this.cantInfo = 0;
-	this.entropia = 0;
+	this.cantInfo = (double) 0;
+	this.entropia =(double) 0;
 }
 
 @Override
